# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API ADM GO",
        "description": " API de aplicación ADM GO.",
        "contact": {
            "email": "basesinteligentes@birobid.com"
        },
        "version": "1.0"
    },
    "servers": [
        {
            "url": "http://181.198.213.18:8000",
            "description": "Server Pro"
        }
    ],
    "paths": {
...
        "/api/deuda": {
            "get": {
                "tags": [
                    "Deudas"
                ],
                "summary": "Retorna registros de ADMDEUDA donde el saldo > 0, el ESTADO != null, y el TIPO in ('FAC','NVT','NDB')",
                "operationId": "App\\Http\\Controllers\\DeudaController::GetDeudas",
                "responses": {
                    "200": {
                        "description": "Retorna los registros de ADMDEUDA"
                    },
                    "default": {
                        "description": "Ha ocurrido un error."
                    }
                }
            }
        },
...
        "/api/deuda/{cliente}": {
            "get": {
                "tags": [
                    "Deudas"
                ],
                "summary": "Retorna registros de ADMDEUDA por cliente, donde el saldo > 0, el ESTADO != null, y el TIPO in ('FAC','NVT','NDB')",
                "operationId": "App\\Http\\Controllers\\DeudaController::GetDeudaXCliente",
                "parameters": [
                    {
                        "name": "cliente",
                        "in": "path",
                        "description": "cliente",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Retorna los registros de ADMDEUDA por cliente."
                    },
                    "default": {
                        "description": "Ha ocurrido un error."
                    }
                }
            }
        },
        "/api/deudacuotas/{secuencial}": {
            "get": {
                "tags": [
                    "Deudas"
                ],
                "summary": "Retorna registros la DEUDA y CUOTAS del secuencial dado.",
                "operationId": "App\\Http\\Controllers\\DeudaCuotasController::GetDeudaCuotas",
                "parameters": [
                    {
                        "name": "secuencial",
                        "in": "path",
                        "description": "secuencial",
                        "required": true,
                        "schema": {
                            "type": "number"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Retorna los registros: cabecera => ADMDEUDA, detalles => ADMDEUDACUOTA donde el saldo > 0."
                    },
                    "default": {
                        "description": "Ha ocurrido un error."
                    }
                }
            }
        },
        "/api/deudapos": {
            "get": {
                "tags": [
                    "Deudas"
                ],
                "summary": "Retorna registros de ADMDEUDAPOS, donde el saldo > 0, el ESTADO != null, y el TIPO in ('FAC','NVT','NDB','CHP')",
                "operationId": "App\\Http\\Controllers\\DeudaPosController::GetDeudas",
                "responses": {
                    "200": {
                        "description": "Retorna los registros de ADMDEUDA."
                    },
                    "default": {
                        "description": "Ha ocurrido un error."
                    }
                }
            }
        },
        "/api/deudapos/{cliente}": {
            "get": {
                "tags": [
                    "Deudas"
                ],
                "summary": "Retorna registros de ADMDEUDAPOS por Cliente, donde el saldo > 0, el ESTADO != null, y el TIPO in ('FAC','NVT','NDB','CHP')",
                "operationId": "App\\Http\\Controllers\\DeudaPosController::GetDeudaXCliente",
                "parameters": [
                    {
                        "name": "cliente",
                        "in": "path",
                        "description": "cliente",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Retorna los registros de ADMDEUDAPOS y los Cheques de ADMDEUDA por cliente."
                    },
                    "default": {
                        "description": "Ha ocurrido un error."
                    }
                }
            }
        },
        "/api/deudatotal/{vendedor}": {
            "get": {
                "tags": [
                    "Deudas"
                ],
                "summary": "Retorna la Sumatoria de SALDO en ADMDEUDA del Vendedor indicado, y el TIPO in ('FAC','NVT','NDB').",
                "operationId": "App\\Http\\Controllers\\DeudaTotalXVendedor::DeudaTotal",
                "parameters": [
                    {
                        "name": "vendedor",
                        "in": "path",
                        "description": "vendedor",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Retorna los registros de ADMDEUDACUOTADET con pago indicado."
                    },
                    "default": {
                        "description": "Ha ocurrido un error."
                    }
                }
            }
        },
        "/api/deudacuotasdet/{pago}": {
            "get": {
                "tags": [
                    "Deudas"
                ],
                "summary": "Retorna registros de ADMDEUDACUOTADET afectado por el pago indicado.",
                "operationId": "App\\Http\\Controllers\\DeudasCoutasDetController::CuotasAfectadas",
                "parameters": [
                    {
                        "name": "pago",
                        "in": "path",
                        "description": "pago",
                        "required": true,
                        "schema": {
                            "type": "number"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Retorna los registros de ADMDEUDACUOTADET con pago indicado."
                    },
                    "default": {
                        "description": "Ha ocurrido un error."
                    }
                }
            }
        },
        "/api/detguiacob/{guia}/{secuencial}": {
            "get": {
                "tags": [
                    "Guia Cobro"
                ],
                "summary": "Retorna el 'detalle' de la Guia de Cobro indicada dependiendo el secuencial",
                "operationId": "App\\Http\\Controllers\\DeudasCoutasDetController::ItemDetGuia",
                "parameters": [
                    {
                        "name": "guia",
                        "in": "path",
                        "description": "guia",
                        "required": true,
                        "schema": {
                            "type": "number"
                        }
                    },
                    {
                        "name": "secuencial",
                        "in": "path",
                        "description": "secuencial",
                        "required": true,
                        "schema": {
                            "type": "number"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Retorna el registro de ADMDETGUIACOB con el SECUENCIAL indicado perteneciente al numero de guia."
                    },
                    "default": {
                        "description": "Ha ocurrido un error."
                    }
                }
            }
        }
    },
    "security": [
        []
    ]
}